requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.0.4
asyncio==3.4.3
awesome-slugify==1.6.5
awsebcli==3.12.1
beautifulsoup4==4.4.1
billiard==3.3.0.23
blessed==1.9.5
boto==2.38.0
boto3==1.1.4
botocore==1.2.11
cairocffi==0.8.0
CairoSVG==2.1.3
celery==3.1.17
cement==2.8.2
cffi==1.11.2
click==6.6
colorama==0.3.7
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.8.3
django-appconf==1.0.1
django-celery==3.1.17
django-compressor==2.0
django-crontab==0.7.0
django-datatables-view==1.12.1
django-enumfield==1.2.1
django-smtp-ssl==1.0
django-storages-redux==1.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
Flask==0.11.1
futures==2.2.0
html5lib==1.0b8
httplib2==0.10.3
icalendar==3.10
itsdangerous==0.24
jdcal==1.0.1
Jinja2==2.8
jmespath==0.9.0
kombu==3.0.37
libsass==0.10.1
MarkupSafe==0.23
mock==1.3.0
mysql-connector==2.1.4
mysqlclient==1.3.6
openpyxl==2.2.6
pathspec==0.5.0
pbr==1.8.0
pdfrw==0.4
Pillow==3.0.0
pycparser==2.18
pyparsing==2.1.0
PyPDF2==1.26.0
Pyphen==0.9.4
pypng==0.0.18
pyScss==1.3.4
python-dateutil==2.4.2
python-engineio==1.0.3
python-socketio==1.6.0
pytz==2015.6
PyYAML==3.11
qrcode==5.3
rcssmin==1.0.6
redis==2.10.3
regex==2016.4.15
reportlab==3.4.0
requests==2.6.2
rjsmin==1.0.12
semantic-version==2.5.0
six==1.9.0
sockjs-tornado==1.0.2
SwampDragon==0.4.2.2
tabulate==0.7.5
termcolor==1.1.0
texttable==0.8.4
tinycss2==0.6.1
tornado==4.2.1
tornado-redis==2.4.18
tzlocal==1.2
Unidecode==0.4.19
wcwidth==0.1.6
WeasyPrint==0.41
webencodings==0.5.1
websocket-client==0.35.0
websockets==3.2
Werkzeug==0.11.11
xhtml2pdf==0.2b1
zope.interface==4.1.3
html2text==2018.1.9